Remove the need to restart Nautilus on config change
This adds a "config file change watcher" to the extension.

At first it was based on native Python threads - but the sleep only
awoke when Nautilus itself was active). Now it is based on GLib timers
- far more reliable.

EVENTUALLY I'd like to see it based on some kind of incoming event so
my code doesn't have to poll - but at the moment, that's not the case.