Start implementation - UI and schema
NOT YET FUNCTIONAL

As with mimetypes and filetypes, path_ paterns is an array of patterns
for which an item will be displayed, with negation possible (i.e. must
not match a pattern).

I'm proposing two basic pattern syntaxes - globs and regular
expressions - with the standard "!" prefix for negation.
Encapsulating the expression with "/" will ark it as an RE.

This commit adds the feature to schema and implements the
first pass at the UI - using JS REs and minimatch globs to check
the validity of the strings entered.

Of course the implementation will be python - as yet I do not have
a good feel for the compatibility of python and JS in these areas -
that comes next.

One annoyance is that, when an error message for an array item
is displayed, the "bin" recentered - so back to the drawing board
on THAT then.