Mitigate validation error on load
The new min/max validator did SOMETHING bad to the editor when
run upon initial config load. I suspect it was to do with the finding of
the associate min/max property editor in order to compare property
values.

This code, then, prevents that validator being run until the editor is
fully loaded and ready.

I'm not yet sure how this affects action creation once the editor is
properly loaded - it APPEARS to work, but there are more tests to do.