More styling
To compact the basic information on Command cards more, we need the
"basic" tab of an object with format "catagories" to be able to be
styled "grid-strict". Unfortunately, that isn't possible with JE at the
moment.

Consequently here we break the UI away from the backend model a bit -
because by introducing a "Basic" object to hold that basic information
we CAN format that object as "grid-strict" - but that' not desirable for
the extension itself.

Consequently, we transform the backend config when we read it, by
moving all the non-array properties into an object property named
"Basic" - and when saving we move them all back to where the extension
expects them to be.

That does mean the JSON schema doesn't actually rpresent the backend
config now - it represents the format used in the UI.

The rest of this commit is tweaking the formatting so that the UI still
behaves the same as before and, still looks roughly the same.