Implement clipboard vs selection
The original implementation via xclip was using the default clipboard -
XA_PRIMARY - which is actually the current _selection_. The original
gtk implementation was using SELECTION_CLIPBOARD - the content of
the clipboard - i.e. the copied selection.

To clear things up and make them both usable, I have left %v as the
selection in both cases, and implemented %V as the copied selection.