Implement show-if-true
A new field that defines a command to run in the current shell
to determine if an itel should be shown.

This can be VERY "expensive" in terms of performance, so it is
suggested that other citeria (such as file typ and name patterns)
be used to prune as much as possible the set of candidates for which the command will be executed.

No doc yet