Another attempt at improving interpolation
So, now we quote tokens that will be passed to the shell and escape quotes within those tokens.
We also do a better job of escaping interpolated values within tokens. The effect is not quite
as pleasant as before - we lose the "escaped spaces" feature that allowed us to manipulate the
order of tokens that apply to a specific selection iteration. But it DOES correctly expose shell
features that might be found within the token. May not filly work at the moment.